Bug-fixes
There's still the problem of handling dependencies while emptying. For example,
if node A wants to empty, and its children B and C are being decompressed and
sorted respectively, then there's a problem. B and C cannot be added into the
empty-queue (because a node can belong to only one queue at a time), but A must
be blocked in some way.