more() function
Use a more() function in each Slave to check if the queue has any waiting
nodes; it returns true if there are buffers waiting and false otherwise. The
empty() function is now changed to return true if the queue is empty (!more())
and the threads aren't working on any buffer (getNumberOfSleepingThreads() ==
numThreads_).