Changes to fix kmer bugs
1. Move splitting code to execute synchronously with emptyBuffer()
2. Since splitting code can be called on multiple children of a node, add
protection while accessing children vector
3. Add a mutex to protect buffer. This is unavoidable for cases where
compression has begun and cannot be cancelled. In such cases attempts
to decompress at the same time will be prevented. This is now necessary
because we have separate compressor and decompressor threads.
4. Add in a semaphore to count the number of decompressed nodes. wait() is
called before decompressing and post() is called after merging. This is not
ideal as nodes could build up on the compression queue.