Lists vector
Make access to vector mutually exclusive. This allows copying into a buffer
when compression or decompression is going on. This is enabled by adding a
lists_copy() function which returns a copy of the lists vector to be used
during compression/decompression. The actual vector may be modified
concurrently. This cleans things up and allows copyIntoBuffer() to proceed
without waiting for previously scheduled compressions to complete, for example.